examples/from_zero_to_one/chapter_04/pyproject.toml
[project]
name = "agent-forge-chapter-04"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "openai>=2.7.1",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.11.0",
  "python-dotenv>=1.2.1",
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
